ast_results/notriddle_budget-envelopes/src/com/notriddle/budget/WidgetService.java
// isComment
package com.notriddle.budget;

import android.appwidget.AppWidgetManager;
import android.appwidget.AppWidgetProvider;
import android.content.Context;
import android.content.Intent;
import android.database.ContentObserver;
import android.database.Cursor;
import android.database.sqlite.SQLiteDatabase;
import android.net.Uri;
import android.os.Bundle;
import android.os.Handler;
import android.os.Looper;
import android.util.Log;
import android.widget.RemoteViews;
import android.widget.RemoteViewsService;

public class isClassOrIsInterface extends RemoteViewsService {

    @Override
    public RemoteViewsFactory isMethod(Intent isParameter) {
        final int isVariable = isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, -isIntegerConstant);
        return new RemoteViewsFactory() {

            Cursor isVariable;

            SQLiteDatabase isVariable;

            ContentObserver isVariable;

            public void isMethod() {
                isNameExpr.isMethod("isStringConstant", "isStringConstant");
                isNameExpr = (new EnvelopesOpenHelper(isNameExpr.this)).isMethod();
                isNameExpr = isNameExpr.isMethod("isStringConstant", null);
                isNameExpr.isMethod(isMethod(), isNameExpr.isFieldAccessExpr);
                isNameExpr = new ContentObserver(new Handler(isNameExpr.isMethod())) {

                    public void isMethod(boolean isParameter, Uri isParameter) {
                        isMethod(isNameExpr);
                    }

                    public void isMethod(boolean isParameter) {
                        isNameExpr.isMethod("isStringConstant", "isStringConstant");
                        isNameExpr.isMethod(isNameExpr.this).isMethod(isNameExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                    }
                };
                isNameExpr.isMethod(isNameExpr);
            }

            public void isMethod() {
                isNameExpr.isMethod("isStringConstant", "isStringConstant");
                isNameExpr.isMethod(isNameExpr);
                isNameExpr.isMethod();
                isNameExpr = null;
                isNameExpr.isMethod();
                isNameExpr = null;
            }

            public boolean isMethod() {
                return true;
            }

            public long isMethod(int isParameter) {
                isNameExpr.isMethod(isNameExpr);
                return isNameExpr.isMethod(isIntegerConstant);
            }

            public int isMethod() {
                int isVariable = isNameExpr.isMethod();
                isNameExpr.isMethod("isStringConstant", "isStringConstant" + isNameExpr);
                return isNameExpr;
            }

            public RemoteViews isMethod() {
                return new RemoteViews(isMethod(), isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
            }

            public RemoteViews isMethod(int isParameter) {
                isNameExpr.isMethod("isStringConstant", "isStringConstant" + isNameExpr + "isStringConstant");
                RemoteViews isVariable = new RemoteViews(isMethod(), isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                isNameExpr.isMethod(isNameExpr);
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isMethod(isIntegerConstant));
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isMethod(isNameExpr.isMethod(isIntegerConstant)));
                Intent isVariable = new Intent();
                isNameExpr.isMethod("isStringConstant", (int) isMethod(isNameExpr));
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr);
                int isVariable = isNameExpr.isMethod(isIntegerConstant);
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, "isStringConstant", isNameExpr);
                return isNameExpr;
            }

            public int isMethod() {
                return isIntegerConstant;
            }

            public void isMethod() {
                isNameExpr.isMethod("isStringConstant", "isStringConstant");
                isMethod();
                isMethod();
            }
        };
    }
}
